Make Filterbank.Header read! work with IOBuffer
Base.read!(::IO, ::Filterbank.Header) used to use `filesize()` on the IO
value passed in, butathe  `filesize()` function passed it's argument(s?)
to `joinpath()` which only takes one or more AbstractString values
(i.e. file names), which IOBuffer clearly is not.  read!() now seeks to
the end of the IO value and queries that end position to figure out the
overall size.