Add Base.size(::Header) for GuppiRaw/Filterbank
The Base.size() function can now be passwd a GuppiRaw.Header or a
Filterbank.Header object.  The returned tuple describes the dimensions
of the data array correspinding to the passed Header object.

The Base.Array() methods in these two modules have been changed to use
the size method for determining the size of the associated file data.
The Array() methods still take additional parameters to influence the
size/dimensionality of the returned Array.