Add Filterbank.mmap functionality
This allows one to work with a filterbank file using memory mapped IO,
which allows one to have an Array spanning more data than can fit in the
system's memory.  It does this by paging in the data from disk as needed
and evicting "old" data pages to make room for "new" data pages as
required.

This uses Julia's `Mmap.mmap` function (which uses the `mmap` system
call).  In order to work natively with 32 bit data values, the data in
the file must have the proper endianness and must be aligned on a four
byte boundary.  This latter requirement is often not met since the
Filterbank header is variable length depending on the lengths of various
strings that are stored in the header.  It would be preferable for data
strings to be NUL padded to the next multiple of 4, but this is not a
strict requirement of the Sigproc/Filterbank specification.  Other
solutions may be available and preferable.

To work with mis-aligned data, the data portion can be mapped as a UInt8
Array with an additional leading dimension of 4.  This can then be
treated as a Float32 Array using Julia 1.6's `reinterpret(reshape,
Float32, data)` function.  How to work around this pre-1.6 remains an
outstanding question.