Improve mmap type alignment warning message
Filterbank files whose data array is not type aligned (e.g. Float32 data
with a header size that is not a multiple of 4) can still be mmap'd, but
it is not so efficient as when the data array is type aligned.  The
warning message printed for type mis-aligned data now includes the
Filterbank file name and does not include the Filterbank.jl source code
location info for where the warning message is generated.