Use weakdeps/extensions when available
If weak dependency extensions are available (Julia >= 1.9.0), use them
to provide dynamic inclusion of DataFrames-related functionality.  For
this to work, the project that has Blio as a dependency must also have
DataFrames as a dependency.  It will not work if DataFrames is loaded
from a different environment.

For Julia versions < 1.9.0, the DataFrames-related functionality is
dynamically included by `Requires.jl` when both DataFrames and Blio are
loaded.