GuppiRaw: Base.size->blocksize, ntime->getntime
This commit renames a couple of methods:

`Base.size(grh::GuppiRaw.Header)` is now
`Blio.blocksize(grh::GuppiRaw.Header)`.  The new name is more
descriptive and the old usage was not really true to the meaning of
Base.size (i.e. the methods returns the size of the data block rather
than the size of the header itself).

`Blio.ntime(grh::GuppiRaw.Header)` is now
`Blio.getntime(grh::GuppiRaw.Header)`. . The old name was too similar to
the name of a variable that one might wish to use to store the returned
value.

Both old methods are deprecated.