Restructure modules for independence [breaking]
This BREAKING CHANGE restructures the `Blio`, `GuppRaw`, and
`Filterbank` modules so that all functions that were in `Blio` Care now
moved into `GuppiRaw` and/or `Filterbank`.  This makes `GuppiRaw` and
`Filterbank` more self-contained an separable from one another so that
they could be broken out into separate packages in the future if
desired.  `GuppiRaw` and `Filterbank` now import the `Base` functions
that they add methods to, which simplifies the syntax of the method
definitions.