Data type of 8 bit Filterbank data depends on nifs
If nifs is 1, the data are assumed to be Stokes I only, so they are
treated as UInt8. . If nifs is not 1, the data are assumed to be full
Stokes or cross-pol, so they are treated as Int8.

I'm not quite sure how valid it is to convert full Stokes or cross-pol
data to 8 bits given the different dynamic ranges of I/XX/YY vs
Q/U/V/XY, but the new approach here will likely lead to fewer surprises
than the old approach.