Add optional predicate argument to GuppiRaw.load
The optional predicate argument is a function that will be passed the
Header of the first GuppiRaw block (of each file) and returns `true`
which will allow loading of the file to proceed or `false` which will
prevent the loading. The `predicate` function can examine the `Header`
to determine whether the file is considered valid (e.g. has
expected sizing and/or frequency attributes).